update pkg resources
