+++ package.json
@@ -55,7 +55,7 @@
     "@electron-toolkit/preload": "^3.0.2",
     "@electron-toolkit/tsconfig": "^1.0.1",
     "@electron-toolkit/utils": "^3.0.0",
-    "@electron/notarize": "^3.0.1",
+    "@electron/notarize": "^2.5.0",
     "@rushstack/eslint-patch": "^1.12.0",
     "@vitejs/plugin-vue": "^5.2.4",
     "@vitest/coverage-v8": "^3.2.4",
