+++ src/main/getCorePath.ts
@@ -1,0 +1,32 @@
+import { spawnSync } from 'child_process'
+import path from 'path'
+import { app } from 'electron'
+
+const FINAL2X_CORE_NAME = 'Final2x-core'
+const FINAL2X_CORE_PATH = '/resources/Final2x-core/Final2x-core'
+
+/**
+ * 获取 Final2x-core 的路径
+ * dev模式下，存放在项目根目录下的 resources
+ * 在 electron-builder 中配置 extraResources，打包时将它放入应用程序根目录的 resources
+ * @returns {string} Final2x-core 的路径
+ */
+export function GetCorePath(): string {
+  if (!CheckPipPackage()) {
+    if (process.env.NODE_ENV === 'development') {
+      return path.join(app.getAppPath(), FINAL2X_CORE_PATH)
+    } else {
+      return path.join(path.dirname(process.execPath), FINAL2X_CORE_PATH)
+    }
+  } else {
+    return FINAL2X_CORE_NAME
+  }
+}
+
+function CheckPipPackage(): boolean {
+  const command = FINAL2X_CORE_NAME + ' -h'
+
+  const result = spawnSync(command, { shell: true, encoding: 'utf-8' })
+
+  return result.status === 0
+}
