+++ src/main/index.ts
@@ -1,11 +1,11 @@
 import { app, BrowserWindow, ipcMain, Menu, nativeImage, shell, Tray } from 'electron'
 import { join } from 'path'
 import { electronApp, is, optimizer } from '@electron-toolkit/utils'
-import { KillCommand, RunCommand } from './RunCommand'
+import { KillCommand, RunCommand } from './runCommand'
 import { openDirectory } from './openDirectory'
 
-const icon = join(__dirname, '../../resources/icon.png')
-const trayIcon = join(__dirname, '../../resources/tray.png')
+import icon from '../../resources/icon.png?asset'
+import trayIcon from '../../resources/tray.png?asset'
 
 function createWindow(): void {
   // Create the browser window.
@@ -139,5 +139,5 @@ app.on('window-all-closed', () => {
   }
 })
 
-// In this file you can include the rest of your app"s specific main process
+// In this file you can include the rest of your app's specific main process
 // code. You can also put them in separate files and require them here.
